custom_airflow/requirements.txt
alembic==1.14.1
astroid==3.3.8
cfgv==3.4.0
colorama==0.4.6
croniter==6.0.0
dill==0.3.9
distlib==0.3.9
filelock==3.17.0
greenlet==3.1.1
identify==2.6.7
iniconfig==2.0.0
isort==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
networkx==3.4.2
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
pylint==3.3.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.38
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.29.2
